% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravcutils.R
\name{apply_rowop}
\alias{apply_rowop}
\title{Applies row-wise operation to a matrix and a vector}
\usage{
apply_rowop(matrix, op, row)
}
\arguments{
\item{matrix}{Matrix}

\item{op}{operation to apply. Use backquotes for operators: \code{`+`}.}

\item{row}{vector treated as a single row.}
}
\description{
\code{apply_rowop} applies operation \code{op} to each row of the matrix as
its first argument and to the vector (treated as a row) as its second
argument.
}
