% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravcutils.R
\name{count_by_group}
\alias{count_by_group}
\title{Finds the count of each observation value (assuming discrete values)}
\usage{
count_by_group(obs)
}
\arguments{
\item{obs}{vector of discrete observations}
}
\value{
The list of observations counted by group. Each list item is the list
    with (at least) fields \code{value} and \code{count}.
}
\description{
Finds the count of each observation value (assuming discrete values)
}
