% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravcutils.R
\name{dunn_index}
\alias{dunn_index}
\title{Dunn index to select cluster numbers}
\usage{
dunn_index(data, classification, centroids, stdev)
}
\arguments{
\item{data}{matrix of observations}

\item{classification}{vector of classiffication labels for each observation}

\item{centroids}{Matrix whose each row is the centroid of a cluster}

\item{stdev}{vector of standard deviations for each variable}
}
\description{
Optimal selection is done for the number of clusters that maximize the index
}
