% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravcutils.R
\name{from_excel_date_number}
\alias{from_excel_date_number}
\title{Converts Excel date serial number to POSIXct}
\usage{
from_excel_date_number(x)
}
\arguments{
\item{x}{Date serial number (from Excel)}
}
\value{
POSIX date/time
}
\description{
When date/time field is loaded from Excel it is often in numeric format.
This function converts it to POSIX date/time.
}
