% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravcutils.R
\name{from_pca_to_original}
\alias{from_pca_to_original}
\title{Converts PCA-space point to original variables space}
\usage{
from_pca_to_original(pca, x)
}
\arguments{
\item{pca}{is an object returned by \code{\link[stats]{prcomp}}}

\item{x}{is a vector or a matrix. It can have less components than full PCA
transformation.}
}
\value{
Returns a point in the original space
}
\description{
\code{from_pca_to_original} converts observations given in (latent) PCA-space
into their original space, taking into account scaling and centering.
}
