% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravcutils.R
\name{n_pca_var_level}
\alias{n_pca_var_level}
\title{Computes number of PCA components needed to explain variance}
\usage{
n_pca_var_level(pca, var_level = 0.95)
}
\arguments{
\item{pca}{Is an object returned by \code{\link{stats}{prcomp}}}

\item{var_level}{Level of variance.}
}
\value{
Number of components that explain at least \code{var_level} of data
    variability
}
\description{
\code{n_pca_var_level} computes number of PCA components needed to explain
variance not below the specified level.
}
\examples{
x <- rnorm(100)
y <- rgamma(100, shape = 0.5)
z <- rnorm(100, sd = 2.0)
x1 <- x + y
x2 <- x - 2 * y + z
x3 <- y + z
x4 <- x + y + z
A <- cbind(x1, x2, x3, x4)
pca <- prcomp(A, scale. = TRUE)
n_pca_var_level(pca)
}
