% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravcutils.R
\name{ravc_palette}
\alias{ravc_palette}
\title{Returns a palette with pleasant colors for plots}
\usage{
ravc_palette()
}
\value{
Vector of 6 colors
}
\description{
Returns a palette with 6 colors to use for plotting.
}
