% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravcutils.R
\docType{package}
\name{ravcutils}
\alias{ravcutils}
\title{ravcutils: A collection of utils for data processing automation.}
\description{
A collection of useful functions to automate data processing and
aid ML application for processing industry applications.
}
