% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravcutils.R
\name{rm_NA_vars}
\alias{rm_NA_vars}
\title{Removes variables from dataframe with too many missing values}
\usage{
rm_NA_vars(df, threshold = 0.3)
}
\arguments{
\item{df}{A dataframe (or a list).}

\item{threshold}{Portion of missing values above which a variable will be
excluded from resulting dataframe.}
}
\value{
A dataframe or a list (depending on \code{df} argument).
}
\description{
\code{rm_NA_vars} returns a dataframe with only those variables whose portion
of missing values is below the specified threshold.
}
\examples{
df <- data.frame(a = c(1, NA, NA, 2), b = c(10, NA, 20, 30))
rm_NA_vars(df)
rm_NA_vars(df, threshold = 0.1)
}
