% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravcutils.R
\name{rm_nonvar_vars}
\alias{rm_nonvar_vars}
\title{Removes numeric variables that vary too little}
\usage{
rm_nonvar_vars(df, threshold = 0.001, scale = NA)
}
\arguments{
\item{df}{A dataframe (or a list)}

\item{threshold}{A scalar, standard deviation threshold for excluding variables}

\item{scale}{A scalar or a vector, the value by which to scale standard
deviation. Use \code{NA} to use mean for scale. If scalar, this scale
value applies to all variables. Otherwise, the vector must be at least
as long as the number of variables in the dataframe.}
}
\value{
A dataframe with non-varying numeric variables removed.
}
\description{
\code{rm_nonvar_vars} returns a dataframe with only those variables from the
original dataframe that are either non-numeric or have relative standard
deviation above the specified threshold.
}
\examples{
df <- data.frame(a = c(1.001, 0.9998, 1, 1.00003), b = c(10., 14., 2., 20.),
                 f = c("yes", "yes", "yes", "yes"))
rm_nonvar_vars(df, threshold = 1e-2)
}
