% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravcutils.R
\name{standardize}
\alias{standardize}
\title{Standardizes numerical matrix}
\usage{
standardize(M)
}
\arguments{
\item{M}{Matrix to standardize}
}
\value{
Matrix with each column having mean zero and variance one.
}
\description{
Standardizes numerical matrix
}
