% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravcutils.R
\name{vol_dist}
\alias{vol_dist}
\title{Computes "volume" of distribution}
\usage{
vol_dist(cov.)
}
\arguments{
\item{cov.}{Covariance matrix of the distribution.}
}
\value{
"Volume" of the ellipsoid.
}
\description{
\code{vol_dist} gives an idea of the spread of distribution by computing its
"volume". Distribution is assumed to be Gaussian with covariance matrix
\code{cov.}. Then the volume of the n-dimensional ellipsoid, defined by this
covariance matrix is computed. This function may be useful for clustering to
assess the size of the cluster.
}
